### 2024-05-24 #6779 @editorialbot
[PRE REVIEW]: pDEMtools: conveniently search, download, and process ArcticDEM and REMA products
**Submitting author:** @trchudley (<a href="http://orcid.org/0000-0001-8547-1132">Thomas Russell Chudley</a>)
**Repository:** https://github.com/trchudley/pdemtools
**Branch with paper.md** (empty if default branch):
**Version:** v0.8.0
