### 2024-05-31 cba0429
adapt terrain scripts for flow-aligned hillshade
