### 2024-06-08 8794ef0
flow-aware hillshade modifications
