### 2024-06-09 841e03e
correct text errors in readme/notebook
