### 2024-06-10 #6779 @trchudley
Thanks @AdamRJensen for agreeing to edit. For your information, I head to Greenland on fieldwork next Tuesday 18th June and will be out of all comms contact until Weds 17th July. RE: searching for reviewers, this might not be an uncommon situation for others in my field as well at this time of year! I'm happy to continue to suggest reviewers if necessary.
