### 2024-06-11 #6779 @mn5hk
> 👋 @ArcticSnow, @mankoff & @mn5hk, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi @AdamRJensen, thanks for the invitation. I would like to contribute as a reviewer to this subscription.

Do you have a preferred date by which you would expect an answer from me?
