### 2024-06-13 b6d67ad
fix mosaic load bug on windows
