+++ src/pdemtools/__init__.py
@@ -6,6 +6,6 @@ from ._index_search import search
 
 from ._accessor import DemAccessor
 
-__version__ = "0.8.1"
+__version__ = "0.8.2"
 
 __all__ = ["search", "DemAccessor"]
