+++ src/pdemtools/load.py
@@ -219,8 +219,8 @@ def from_id(
 
     if preview == True:
         browser_prefix = "https://polargeospatialcenter.github.io/stac-browser/#/external/pgc-opendata-dems.s3.us-west-2.amazonaws.com"
-        preview_fpath = os.path.join(
-            browser_prefix, dataset, "strips", version, "2m", geocell, f"{dem_id}.json"
+        preview_fpath = (
+            f"{browser_prefix}/{dataset}/strips/{version}/2m/{geocell}/{dem_id}.json"
         )
         return preview_fpath
 
@@ -396,6 +396,4 @@ def _aws_link(
         raise ValueError(f"Input `resolution` must be one of ['2m', '10m', '32m']")
 
     # Return appropriate filepath.
-    return os.path.join(
-        prefix, dataset, "mosaics", version, resolution, row.supertile, fname
-    )
+    return f"{prefix}/{dataset}/mosaics/{version}/{resolution}/{row.supertile}/{fname}"
