### 2024-07-22 788e53c
fixed index search for new PGC files
