+++ src/pdemtools/__init__.py
@@ -6,6 +6,6 @@ from ._index_search import search
 
 from ._accessor import DemAccessor
 
-__version__ = "0.8.2"
+__version__ = "0.8.3"
 
 __all__ = ["search", "DemAccessor"]
