### 2024-07-28 #6779 @AdamRJensen
Personally, I think JOSS is doing very important work on advancing open science and I hope you'll be able to help out with this mission! 🥳
