### 2024-07-29 4e414f7
address mosaic opening error. fixes #13.
