+++ environment.yml
@@ -14,3 +14,4 @@ dependencies:
   - numpy
   - shapely
   - pyarrow
+  - fiona
