### 2024-07-29 6ae57e8
increase memory efficiency of coreg/geomorph
