+++ batch/batch_download_and_coregister.py
@@ -5,21 +5,30 @@ the current working directory.
 
 The script will download the 2 m ArcticDEM or REMA mosaic for coregistration, and then
 loop through the strip record, downloading and coregistering against stable ground
-identified within the BedMachine mask.
+identified within the BedMachine mask. If coregistration is succesful, the file will 
+be saved ending *_coreg.tif. If coregistration fails (most usually because no stable
+bedrock is available in the strip), the filename will not have `_coreg` appended.
 
 Note: this will only work in regions of Greenland and Antarctica where bare rock
-is exposed. If you require coregistration in other regions, please drop me an email - 
-I would be happy to help.
+is exposed. If you require coregistration in regions outside of Greenland or Antarctica, please 
+drop me an email - I would be happy to help.
 
 Tom Chudley | thomas.r.chudley@durham.ac.uk
 Durham University
-November 2023
+
+v1 | 2023-11 | Initial script
+v2 | 2024-07 | Updated to make *_coreg.tif fname dependent on sucessful coregistration.
+
 """
 
 import os
+
 import rioxarray as rxr
 import numpy as np
 import pdemtools as pdt
+
+from glob import glob
+from math import isnan
 
 # ----------------------------------------------------------------------------------- #
 # EDIT THIS SECTION TO SPECIFY DEPENDENT FILES, THE REGION, AND BOUNDS OF INTEREST
@@ -68,9 +77,9 @@ if not os.path.exists(reference_dem_fpath):
 
     reference_dem = pdt.load.mosaic(
         dataset=dataset,  # must be `arcticdem` or `rema`
-        resolution=2,  # must be 2, 10, or 32
-        bounds=bounds,  # (xmin, ymin, xmax, ymax) or shapely geometry
-        version="v4.1",  # optional: desired version (defaults to most recent)
+        resolution=2,     # must be 2, 10, or 32
+        bounds=bounds,    # (xmin, ymin, xmax, ymax) or shapely geometry
+        version="v4.1",   # optional: desired version (defaults to most recent)
     )
     reference_dem.rio.to_raster(
         reference_dem_fpath, compress="ZSTD", predictor=3, zlevel=1
@@ -104,7 +113,7 @@ n_strips = len(gdf)
 
 print(f"{n_strips} strips found")
 
-i = 0
+i = 1
 
 print("\nDownloading DEM strips...")
 for _, row in gdf.iterrows():
@@ -112,14 +121,34 @@ for _, row in gdf.iterrows():
     date_str = date.strftime("%Y%m%d")
     dem_id = row.dem_id
 
-    out_fpath = os.path.join(outdir, f"{date_str}_{dem_id}_coreg.tif")
+    out_fname = os.path.join(outdir, f"{date_str}_{dem_id}")
 
-    if not os.path.exists(out_fpath):
+    # If the file doesn't yet exist, download.
+    if len(glob(f'{out_fname}*')) == 0:
+
         print(f"\nDownloading {i}/{n_strips} {os.path.basename(out_fpath)}...")
+
+        # download DEM
         dem = pdt.load.from_search(row, bounds=bounds, bitmask=True)
         dem.compute()  # rioxarray uses lazy evaluation, so we can force the download using the `.compute()` function.
+        
+        # pad to full size of AOI
         dem = dem.rio.pad_box(*bounds, constant_values=np.nan)
-        dem = dem.pdt.coregister(reference_dem, bedrock_mask, max_horiz_offset=50)
+
+        # coregister DEM, with return_stats=True.
+        dem = dem.pdt.coregister(reference_dem, bedrock_mask, max_horiz_offset=50, return_stats=True)
+        
[…]
