### 2024-07-29 b9f43f9
update tests to account for newer float32 approach
