### 2024-07-29 7c18d50
update tests to account for newer float32 approach
