+++ tests/test_processing.py
@@ -78,7 +78,7 @@ def test_geomorphometry():
 
     variable_min_max = {
         "slope": (0.0, 24.721811294555664),
-        "aspect": (1.3660377589985728e-05, 359.9999694824219), #(0.0, 360.0),
+        "aspect": (0.0, 360.0), # (1.3660377589985728e-05, 359.9999694824219),
         "hillshade": (0.0, 1.0),
         "horizontal_curvature": (-0.0007999989320524037, 0.0007856095326133072),
         "vertical_curvature": (-0.027519449591636658, 0.022581202909350395),
