### 2024-08-05 #6779 @PennyHow
@AdamRJensen, I have been out of signal for some time, so sorry for the delay in answering.

I would be happy to review, but I am now on a mixture of fieldwork and conferences until mid-September. So I would not be able to start a review until then. Is that okay?
