### 2024-08-27 #7149 @editorialbot
[REVIEW]: pDEMtools: conveniently search, download, and process ArcticDEM and REMA products
**Submitting author:** @trchudley (<a href="http://orcid.org/0000-0001-8547-1132">Thomas Russell Chudley</a>)
**Repository:** https://github.com/trchudley/pdemtools
**Branch with paper.md** (empty if default branch):
**Version:** 0.8.5
