### 2024-08-29 trchudley/pdemtools#16 @mn5hk
Code review for JOSS
I will be jotting down my review notes here, if any, as part of the JOSS review process openjournals/joss-reviews#7149

`installed well with conda on a linux machine`
- [Supplementary Data](https://pdemtools.readthedocs.io/en/latest/getting_started/supplementary_datasets.html):  [BedMachineGreenland-v5.nc](https://nsidc.org/data/idbmg4/versions/5) and [ArcticDEM Strip Index](https://data.pgc.umn.edu/elev/dem/setsm/ArcticDEM/indexes/)
- [`Mosaic and Terrain`](https://pdemtools.readthedocs.io/en/latest/examples/mosaic_and_terrain.html) works as described.
- [`Strip search and coregistraion`](https://pdemtools.readthedocs.io/en/latest/examples/strip_search_and_dem_difference.html) works as described; very well documented.
- [`Get Icebergs`](https://pdemtools.readthedocs.io/en/latest/examples/get_icebergs.html)
- minor issue: `ModuleNotFoundError: No module named 'skimage'`
- quick fix: `conda install -c conda-forge scikit-image`
- `Flow-aware Hillshade`: works as described
- Not included in documentation..
