### 2024-08-29 trchudley/pdemtools#16 @mn5hk
Hi @trchudley,

I am currently trying to test the [`Strip search and coregistraion`](https://pdemtools.readthedocs.io/en/latest/examples/strip_search_and_dem_difference.html) script. However, I am currently stuck with an incorrect `ArcticDEM_Strip_Index_s2s041_gpqt` file that I have downloaded online.

Could you possible share a download link to the file?

Thanks,
Amin
