### 2024-08-29 trchudley/pdemtools#16 @trchudley
Hi Amin @mn5hk

No worries, thanks for reviewing!

The ArcticDEM index is [this one](https://data.pgc.umn.edu/elev/dem/setsm/ArcticDEM/indexes/ArcticDEM_Strip_Index_s2s041_gpqt.zip), but it must be unzipped and the resultant .gpqt referred to. If you're having any issues with that, do let me know and I'll try and diagnose.

I link to and try and explain the index dataset in the [Supplementary Datasets](https://pdemtools.readthedocs.io/en/latest/getting_started/supplementary_datasets.html) section of the Documentation, but if there's any way it could be made clearer please do let me know!

Cheers,
Tom
