### 2024-08-29 trchudley/pdemtools#16 @mn5hk
Hi Thomas @trchudley,

Thanks for a quick response. Indeed, I overlooked the `Supplementary Data` section earlier.

The code worked well for me. As of now, I have gone through the example notebooks provided in the repository. They all work well and are well documented. I hope to finish the rest of the review soon.

So far I have only couple minor suggestions:

- [ ] `scikit-image`could be added to the `environment.yml` file. This caused a minor hiccup on the [`get icebergs`](https://github.com/trchudley/pdemtools/blob/main/notebooks/get_icebergs.ipynb) script.
- [ ] `flow_aware_hillshades` is a nice example in the repository. It is not currently accessible in the [documentation website](https://pdemtools.readthedocs.io/)

Best,
Amin
