### 2024-08-29 trchudley/pdemtools#16 @trchudley
Hi Amin @mn5hk

Thanks so much for your review!

I will be sure to update the strip search notebook to reference the `Supplementary Data` page for those encountering it for the first time.

### `scikit-image`

I am aware that `scikit-image` is not included as a dependency. This was a conscious decision made due to the fact that it introduces a lot of further dependencies that would bloat a standalone `conda` environment, despite it being only used for this 'optional' example analysis (i.e. unrelated to DEM processing) in this notebook. One option might be to include it as an extra optional dependency in the pip install (e.g. `pip install pdemtools[skimage]`), but, as we identify on the [Install documentation](https://pdemtools.readthedocs.io/en/latest/getting_started/install.html), `pip` is not a recommended way of installing `pdemtools`, whilst [`conda` has yet to include optional dependencies](https://github.com/conda/conda/issues/7502).

My preferred option here would be to make this explicit in the example notebook - e.g. add:

> `scitkit-image` is not a mandated dependency of `pdemtools`, and must be installed separately (e.g. via `conda install -c conda-forge scikit- `).

Would you agree that this is a suitable approach?

### `flow_aware_hillshades`

Good catch! I have refrained from adding it to the website at this point for two reasons:

 - Officially, I have submitted version 0.8.0 for review, and this notebook requires version 0.8.1. In practice, I know the latest `conda` version will be what is actually reviewed (with some relief, as version 0.8.0 had a bug that meant it didn't work on Windows!), but this will keep up the pretence.
 - I am undecided on whether this notebook is too 'complex', by which I mean there is a lot of additional code that isn't wrapped within `pdemtools` functions. I have made it available here to share with a student who is working on this problem, without 'publicising' it on the website.

My working plan at this point is to add it to the website after the reviews have been completed, but am open to adding it earlier if you think it is still a good idea.
