### 2024-08-30 #7149 @mn5hk
Dear editor @AdamRJensen ,

I am done with my review (of the latest version: v.0.8.4), and I recommend the submission be accepted without any additional modifications. Furthermore, I thank the authors for quick clarifications on my minor queries, which have been linked to this Github issue thread.

I have focused my review primarily on the quality of the code functionality and documentation. The code is well-written, with excellent documentation and some practical examples of how this code may be used. The API documentation and community guidelines are satisfactory. Lastly, I appreciate the extensive unit tests and Git workflow automation incorporated into the repository.

The manuscript is satisfactory, with good references to other alternative toolkits. I appreciate the extensive citation of relevant tools and methods throughout the manuscript, repository, and software documentation.

The software is focused on streamlining polar Digital Elevation Models (DEM) download and preprocessing steps. The examples provided in the documentation are good starting points for new users to familiarize themselves with the software and DEM preprocessing. I am of the opinion that the software can make a good contribution in simplifying such preprocessing steps and aiding further advancement of research in the field.

Lastly, as an Early Career Researcher and an aspiring scientific programmer, I thank the editor and the authors for the opportunity to review this submission. I enjoyed testing the software and hope it will be useful in my own future research activities (and many other researchers' work).

Sincerely,
Amin
