### 2024-10-04 trchudley/pdemtools#17 @maiwinstrup
Issues with supplementary data download 
With regards to the supplementary data that the user is required to download for the script to work, I needed some additional information on exactly which files are required, and where these files should be placed locally. The authors may consider to add convenience for the users by providing automatic download (and unzip) of the required data sets (in the notebooks).

Specifically, I was lacking the following information:
•	Specify exactly which index files are needed to download, e.g. filename and number of files. Does one need to download index files for the mosaic? (I don’t think so?). And should it be the ‘latest’ version, or the files with a specific version? (These are currently identical, but this is not likely to stay this way).
•	Also, will the code provided in the notebooks be able to cope when a new version of index files becomes the ‘latest’?
•	Same for bedmachine data – specify that one needs to download the .nc file, not the _bed.tif file
•	Mention which folder these files should be extracted to for the notebooks to run without adjustments

openjournals/joss-reviews#7149
