### 2024-10-04 trchudley/pdemtools#18 @maiwinstrup
Version number
Update version updates in the documentation to reflect the github version number. openjournals/joss-reviews#7149
