### 2024-10-04 trchudley/pdemtools#19 @maiwinstrup
How large an area is appropriate to work with?
I was lacking was some guidance regarding how big of an area is ‘appropriate’ to work with using the software. I several times ran into memory issues (running it on a new laptop) because I had selected an area that was too large. openjournals/joss-reviews#7149
