### 2024-10-04 trchudley/pdemtools#20 @maiwinstrup
Geoid clarification in the notebooks
As most users will approach the software using the notebooks, it should here be specified which geoid is being used for the geoid correction (so that all users do not have to check which one is used in the bedmachine dataset). openjournals/joss-reviews#7149
