### 2024-10-04 trchudley/pdemtools#17 @maiwinstrup
Also, my ArcticDEM strip index file is not ending with '..._gpqt.parquet' (as the path provided in the notebook), but just '.parquet'
