### 2024-10-04 trchudley/pdemtools#21 @maiwinstrup
bug in pdt.search()?
Please check the following: When running the pdt.search with parameters provided below, incl. min_aoi_frac=0.5, I get ‘5 strips found’, but the last strip (fid=16290) appears to have no data within the area? Why? Is this due to an error in the code?

gdf = pdt.search(
    index_fpath,
    bounds,
    dates = '20160101/20221231',
    months = [6,7,8,9],
    baseline_max_hours = 24,
    sensors=['WV03', 'WV02', 'WV01'],
    min_aoi_frac = 0.5,
)
