### 2024-10-04 trchudley/pdemtools#22 @maiwinstrup
Include aoi_frac in results from pdt.search()
Adding aoi_frac in resulting gdf from pdt.search() would add convenience when judging which of the selected scenes may be best. openjournals/joss-reviews#7149
