### 2024-10-04 trchudley/pdemtools#23 @maiwinstrup
Plots in flow_aware_hillshade notebook
The plots in the flow_aware_hillshade notebook were not showing for me - not sure why. They appeared after adding plt.show() to the plotting code.  openjournals/joss-reviews#7149
