### 2024-10-04 #7149 @maiwinstrup
**Review of pDEMtools (v.0.8.4) by Chudley et al**

The software enables easy download and preprocessing of the high-resolution polar DEMs data sets from stereo photogrammetry (ArcticDEM and REMA). Very nice work, and well-documented. Most of my remarks/issues are minor, and should be easy to fix. They primarily concern making the software (even) more user-friendly.

A potential bug has been submitted as a github issue, along with a few requested clarifications on the code and documentation. Further, there are a few typos in the notebooks and the API documentation (I have not corrected these).

The accompanying software paper is well-written, and explains well the merits of the product. One comment: Please explain the difference between a dynamic STAC and a ‘normal’ STAC (L31), and the abbreviation RESTful (L31).

After fixing these, and accounting for the remarks below, I recommend the acceptance of the submission.

_**Installation:**_
Nice introduction in the ‘readthedocs’, ensuring easy download and install that (mostly) works out of the box. I did, however encounter some issues regarding e.g., clarification on supplementary data download and file location (issue raised on github)

I enjoyed following the examples in the notebooks, which are provided as part of the source code. The same examples are provided as part of the readthedocs. I am, however, missing a reference to the notebooks in the “readthedocs”, to ensure that users reading this document are aware that the notebooks exist.

_**Functionality:**_
I have tested the main functionalities, primarily based on the examples provided in the notebooks. Everything seems to work as it should, except for a potential bug when searching for strip DEMs in areas mostly covered by ocean (issue raised on github).

The flow_aware_hillshade notebook is written in a slightly different style from the other notebooks. I suggest to revise it to better align with the others.

Disclaimer: I have only tested the code for Greenland (ArcticDEM), not for Antarctica.

_**Performance:**_
No claims given
