+++ docs/appendix/version_updates.md
@@ -1,10 +1,11 @@
 # Version updates
 
+From the public conda release of version 0.8.0 onwards, [version updates and notes are automatically recorded on the GitHub repo 'releases' page`](https://github.com/trchudley/pdemtools/releases). 
+
+Versions prior to this are recorded in the table below for posterity:
+
 | Version | Date | Notes |
 | ------- | ---- | ----- |
-| 0.8.2 | June 2024 | Bug fix to address fail loading of mosaics on Windows systems |
-| 0.8.1 | June 2024 | Minor modification to hillshade script to allow for MacGregor _et al._ (2024) flow-aware hillshade. Supplementary notebook available in GitHub `notebooks` directory. |
-| 0.8.0 | May 2024 | Iceberg detection now corrects for sea surface; final additions to documentation before public conda/pip release.
 | 0.7.0 | March 2024 | Minor bug fixes and more detailed docstrings to support readthedocs | 
 | 0.6.0 | February 2024 | Added `data.bedrock_mask_from_vector()` function. |
 | 0.5.0 | January 2024 | Fixed bug in `_coreg.py` to improve Nuth and Kääb (2011)-style coregistration. |
