+++ README.md
@@ -40,16 +40,15 @@ Please visit the [pDEMtools readthedocs](https://pdemtools.readthedocs.io/en/lat
 Update when v.1.0 uploaded to Zendoo.
  -->
 
- The use of the pDEMtools package can be cited as follows:
+A software paper is being prepared for the [Journal of Open Source Software](https://joss.theoj.org/). In the meantime, the use of the pDEMtools package can be cited as follows:
 
-
-> Chudley, T. R. and Howat, I. M. (2024) pDEMtools (vX.X.X). GitHub. https://github.com/trchudley/pDEMtools
+> Chudley, T. R. and Howat, I. M. (2024) pDEMtools: conveniently search, download, and process ArcticDEM and REMA products (vX.X.X). GitHub. https://github.com/trchudley/pDEMtools
 
 or by using `bibtex`:
 
 ```
 @software{pDEMtools
-   author = {Chudley, Thomas R. and Howat, Ian M.}, title = {pDEMtools}, year = 2024, publisher = {GitHub}, version = {X.X.X}, url = {https://github.com/trchudley/pDEMtools} 
+   author = {Chudley, Thomas R. and Howat, Ian M.}, title = {pDEMtools: conveniently search, download, and process ArcticDEM and REMA products}, year = 2024, publisher = {GitHub}, version = {X.X.X}, url = {https://github.com/trchudley/pDEMtools} 
 }
 ```
 
