+++ src/pdemtools/_index_search.py
@@ -25,6 +25,7 @@ def search(
     sensors: Optional[str | tuple] = None,
     is_xtrack: Optional[bool] = None,
     accuracy: Optional[float | tuple] = None,
+    include_custom_columns: Optional[bool] = True,
 ):
     """Efficiently search the ArcticDEM and REMA strip index files provided by the
     Polar Geospatial Center.
@@ -66,6 +67,11 @@ def search(
         value as an upper range (e.g. 2), or a tuple of two values in order to include
         a lower bound and filter NaN values (e.g [0, 2]).
     :type accuracy: float | tuple, optional
+    :param include_custom_columns: Whether to include custom columns generated by
+        `pdemtools` in geopandas dataframe output. This includes "pdt_time1",
+        "pdt_time2", "pdt_dem_baseline_hours", "pdt_time_mean", "pdt_year", "pdt_month",
+        and "pdt_aoi_frac". Defaults to True
+    :type include_custom_columns: bool, optional
 
     :returns: Strip index filtered to desired variables.
     :rtype: GeoDataFrame
@@ -235,19 +241,32 @@ def search(
         gdf["aoi_frac"] = gdf.area / aoi_area
         gdf = gdf[gdf["aoi_frac"] > min_aoi_frac]
 
-    # Drop created columns
-    gdf = gdf.drop(
-        [
-            "time1",
-            "time2",
-            "dem_baseline_hrs",
-            "time_mean",
-            "month",
-            "year",
-            "aoi_frac",
-        ],
-        axis=1,
-        errors="ignore",
-    )
+    # If include_custom_columns is True, rename columns to begin with "pdt_"
+    if include_custom_columns:
+        gdf = gdf.rename(columns={
+            "time1": "pdt_time1", 
+            "time2": "pdt_time2",
+            "dem_baseline_hours": "pdt_dem_baseline_hours",
+            "time_mean": "pdt_time_mean",
+            "year": "pdt_year",
+            "month": "pdt_month",
+            "aoi_frac": "pdt_aoi_frac",
+        })
+
+    # Else, drop the created columns
+    else:
+        gdf = gdf.drop(
+            [
+                "time1",
+                "time2",
+                "dem_baseline_hrs",
+                "time_mean",
+                "month",
+                "year",
+                "aoi_frac",
+            ],
+            axis=1,
+            errors="ignore",
+        )
 
     return gdf
