+++ .github/workflows/unit_test.yml
@@ -14,6 +14,7 @@ jobs:
       matrix:
         python-version:
           - "3.10"
+          - "3.11"
 
     steps:
       - name: Checkout repo
