+++ environment.yml
@@ -3,7 +3,7 @@ channels:
   - conda-forge
   - defaults
 dependencies:
-  - python=>3.10
+  - python>=3.10
   - rioxarray 
   - geopandas 
   - opencv 
