### 2024-10-07 trchudley/pdemtools#17 @trchudley
> The authors may consider to add convenience for the users by providing automatic download (and unzip) of the required data sets (in the notebooks).

An early version of this repo did have bash scripts that provided functionality - however, we found that they were only intermittently succesful, as it looks like this part of the PGC website has some protection against automated data retrieval. As a result, we reverted to explicit instructions for manual retreival so as not to confuse users with inconsistent scripts.

> - Specify exactly which index files are needed to download, e.g. filename and number of files. Does one need to download index files for the mosaic? (I don’t think so?). And should it be the ‘latest’ version, or the files with a specific version? (These are currently identical, but this is not likely to stay this way).

Yes, only a single index file is needed (no mosaics necessary - these are small so they are actually [included within `pdemtools`](https://github.com/trchudley/pdemtools/tree/main/src/pdemtools/mosaic_index)). The `latest_gpkt` version should work fine but you are correct that, because these may drift in the future, we the `s2s041_gpkt` to ensure that documentation remains usable even if the `latest` version is updated without notice.

__Action:__ We have updated the Supplementary Data page to specify this information.

> - Also, will the code provided in the notebooks be able to cope when a new version of index files becomes the ‘latest’?

Ultimately it is the PGC who are responsible for distributing the products and so 'we' (i.e. speaking as `pdemtools` authors) cannot guarantee that the future index files will remain compatible with our software without any active maintenance on our side. However, we have already gone through one relatively large annual update (an annual update of index files and an ArcticDEM mosaic update v3 to v4.1) without an issue. The code worked without issue, I merely updated the links and references in the documentation, and updated the mosaic package in the background.

I have also been in touch with @clairecporter from the PGC who has told me that (i) it is unlikely that the strip version will incremenet again - at least in the near future - and simply more data will be added to the existing files; and (ii) even should things increment, future releases will aim to retain backwards compatibility (e.g. adding new fields rather than changing existing field names), and any breaking changes will be clearly communicated ahead of time, giving us time to adjust should the worst case scenario happen.

__Action:__ On the Supplementary Data page, we will clarify that the software is known to work on the most recent update of the ArcticDEM/REMA datasets (Strip Release 2023, released Aug 2024)

> - Same for bedmachine data – specify that one needs to download the .nc file, not the _bed.tif file

__Action:__ We have updated the Supplementary Data page to clarify this.

> - Mention which folder these files should be extracted to for the notebooks to run without adjustments

The files do not need to be extracted anywhere in particular: instead, the user provides a filepath (e.g. as the `index_fpath` variable in the `pdt.search()` function, or the `bm_fpath` variable within the `pdt.data.bedrock_mask_from_bedmachine()` function). This was a deliberate choice. Many users have set locations for large public datasets within their file system (I know I personally have a `~/data` symlink to my local cluster which contains e.g. BedMachine, ArcticDEM/REMA mosaics, GrIMP mask, ITS_LIVE annual velocity mosaics, QGreenland/Quantarctica data, Greenland/Antarctica shapefiles, etc etc.). Asking users to place the index/BedMachine/their own geoids in a specific directory may mean that they either break other scripts that reference the users preferred location, or the data is duplicated, which can be signficant (e.g. ~2.7 GB for BedMachine Greenland). Asking the user to provide a filepath solves this potential problem, although we acknowledge that specifying a directory for necessary files is another perfectly valid way to approach the issue of required external datasets.

__Action:__ We have clarified this on the Supplementary Data Page.

> Also, my ArcticDEM strip index file is not ending with '..._gpqt.parquet' (as the path provided in the notebook), but just '.parquet'

Thanks for this - this is a good catch. It is the PGC zip file that will end `_qpqt.zip`, not the contained parquet file itself. We will update all references to this.

Hopefully the readthedocs website will update to reflect these updates in the next few hours.
