### 2024-10-07 trchudley/pdemtools#18 @trchudley
__Action:__ We have updated the page to redirect users to the [GitHub releases page](https://github.com/trchudley/pdemtools/releases), which automatically records all version updates since the public pip/conda release at v.0.8.0. This page now only records pre-v0.8.0 changes.

Hopefully the readthedocs website will update to reflect these updates in the next few hours.
