### 2024-10-07 trchudley/pdemtools#20 @trchudley
__Action__ We have clarified in the notebooks and supplementary datasets page that the EIGEN-C64 geoid is the one used by the (current) BedMachine datasets.

Hopefully the readthedocs website will update to reflect these updates in the next few hours.
