### 2024-10-07 trchudley/pdemtools#22 @trchudley
__Action:__ This is a good idea. In fact, this column (among others) is already created as part of the index search process - it is just explicitly dropped before the gdf is returned. We have created a new `include_custom_columns`, set to `True` by default, which returns these custom columns ('["pdt_time1", "pdt_time2", "pdt_dem_baseline_hrs", "pdt_time_mean", "pdt_month", "pdt_year", "pdt_aoi_frac"]`) as part of the gdf so that users can benefit from them.

This functionality is now present in the `main` github repo so is available for developer testing. Upon acceptance to JOSS, we will push the change to a new pip/conda version v0.8.5.
