### 2024-10-07 trchudley/pdemtools#23 @trchudley
Whilst we couldn't replicate this, after doing some research it appears that this could be related to some of the fine points of Jupyter Notebook/Lab and the %matplotlib magic command.

__Action:__ We have now explicitly included `%matplotlib inline` at the top of all notebooks, which should, to the best of knowledge, now enforce things to work - assuming it is being run in a properly installed JupyterLab environment rather than the deprecated 'classic' Jupyter Notebook.
