### 2024-10-07 #7149 @trchudley
Thanks @maiwinstrup for such a detailed and thorough review! We really appreciate the time you've taken on this.

We have outlined our responses below and pushed all the changes to the github. Please note that we would appreciate some clarification for this issue https://github.com/trchudley/pdemtools/issues/21.

The updates to the documentation should filter through to the website within hours. Once @AdamRJensen and yourself are satisfied with our responses and that `pdemtools` is suitable for publication in JOSS, we are prepared to push the minor changes to `pip` and `conda` as a new version (0.8.5) alongside/prior to publication.

## Responses

> The software enables easy download and preprocessing of the high-resolution polar DEMs data sets from stereo photogrammetry (ArcticDEM and REMA). Very nice work, and well-documented. Most of my remarks/issues are minor, and should be easy to fix. They primarily concern making the software (even) more user-friendly.

We are grateful for your kind words and detailed review.

> A potential bug has been submitted as a github issue, along with a few requested clarifications on the code and documentation.

__Action__:

 - We have addressed the potential bug in the GitHub issue (https://github.com/trchudley/pdemtools/issues/21). Please note that we require some clarification on this to properly resolve, although we think we have identified the problem and proposed a potential solution.
 - We have added an `aoi_frac` output to the `search()` output as requested (https://github.com/trchudley/pdemtools/issues/22).
 - We have updated the documentation where clarifications and corrections have been requested (https://github.com/trchudley/pdemtools/issues/18, https://github.com/trchudley/pdemtools/issues/19, https://github.com/trchudley/pdemtools/issues/20)

> Further, there are a few typos in the notebooks and the API documentation (I have not corrected these).

__Action:__ We have gone through the notebooks and attempted to correct typos where visible.

> The accompanying software paper is well-written, and explains well the merits of the product. One comment: Please explain the difference between a dynamic STAC and a ‘normal’ STAC (L31), and the abbreviation RESTful (L31).

The Spatio Temporal Asset Catalog specification is a standardised way of storing geospatial data as a network of JSON files ([see here for a good intro](https://developers.planet.com/docs/planetschool/introduction-to-stac-part-1-an-overview-of-the-specification/)). A 'static' catalog is simply a collection of JSON can be viewed on file servers or the cloud. A dynamic STAC can be queried over the internet in real time, by asking via HTML in a specific way (an API interface known as Representational State Transfer, or REST). Tools such as `pystac` allows the user to ask for, and retreive, data from the dynamic STAC.

__Action:__ We have updated the software paper to clarify the meaning of static and dynamic, and we have removed the reference to a RESTful API as, given the target audience of our software and the intended high-level overview of the manuscript, it is not necessary to understand this complexity here.

> After fixing these, and accounting for the remarks below, I recommend the acceptance of the submission.

> __*Installation:*__

> Nice introduction in the ‘readthedocs’, ensuring easy download and install that (mostly) works out of the box. I did, however encounter some issues regarding e.g., clarification on supplementary data download and file location (issue raised on github).

We have responded to the supplementary data download queries by updating the relevant page in the documentation, described in detail on the issue (https://github.com/trchudley/pdemtools/issues/17).

> I enjoyed following the examples in the notebooks, which are provided as part of the source code. The same examples are provided as part of the readthedocs. I am, however, missing a reference to the notebooks in the “readthedocs”, to ensure that users reading this document are aware that the notebooks exist.

__Action:__ We have added the relevant GitHub link to the top of every notebook, so that they can be found from the readthedocs (and whereever else they end up).

> __*Functionality:*__

> I have tested the main functionalities, primarily based on the examples provided in the notebooks. Everything seems to work as it should, except for a potential bug when searching for strip DEMs in areas mostly covered by ocean (issue raised on github).

As mentioned above, please note that we would appreciate some clarification for this issue https://github.com/trchudley/pdemtools/issues/21.

> The flow_aware_hillshade notebook is written in a slightly different style from the other notebooks. I suggest to revise it to better align with the others.

We have previously discussed the `flow_aware_hillshade` notebook with Reviewer 1 [here](https://github.com/trchudley/pdemtools/issues/16#issuecomment-2317695610). In brief, it is a work-in-progress that I have made it available on the GitHub to share with a student who is working on this problem without 'publicising' it on the website. This may explain why the tone is different, why it uses some further datasets without detailed introduction, and why it introduces complex fucntions that are not wrapped within `pdemtools`.

__Action:__
 - If it is all right, I would to keep this as it is for now and introduce it to the 'public' readthedocs at the appropriate time in the future. However, I have added a note to the top of the notebook clarifying this for users who come across it.
 - I have also attempted to fix the `plt.show()` error as requested (https://github.com/trchudley/pdemtools/issues/23).
