### 2024-10-07 54de1a1
polish notebooks #17
