+++ docs/getting_started/supplementary_datasets.md
@@ -1,6 +1,6 @@
 # Supplementary datasets
 
-To make the most of pDEMtools, two supplementary datasets must available locally.
+To make the most of pDEMtools, two supplementary datasets must be available locally.
 
 ## ArcticDEM/REMA strip index
 
@@ -8,7 +8,7 @@ The first dataset is the ArcticDEM or REMA strip index made available by the PGC
 
 The appriate files to download are the `ArcticDEM_Strip_Index_s2s041_gpqt.zip` and `REMA_Strip_Index_s2s041_gpkg.zip` - note the __GeoParquet file format__ (ending `_gpqt.zip`), *not* the similarly named GeoPackage (ending `_gpkg.zip`). No mosaic files are necessary - these are included with `pdemtools`. 
 
-Unzip the files before before use. They can be placed anywhere in your file system - the filepath, as a string, is provided to the relevant `pdemtools` functions.
+Unzip the files before use. They can be placed anywhere in your file system - the filepath, as a string, is provided to the relevant `pdemtools` functions.
 
 > __NOTE:__ These index files are maintained by the PGC, so we cannot guarantee that newly released updates to ArcticDEM/REMA index files will not temporarily break `pdemtools` functions until we can fix and update the software. From version 0.8.3 onwards, these files are known to work up to and including the latest PGC data update (ArcticDEM Strip Release 2023, released Aug 2024).
 
