### 2024-10-08 trchudley/pdemtools#21 @maiwinstrup
The bounds are: bounds = (-286000, -930400, -282000, -925000) :) It is the very outer part of Petermann, and it has data, when I set bitmask=False, as suggested. Is it possible to plot the data using 'preview' without applying the bitmask? (I tried setting bitmask=False in the pdt.load.preview(), but that wasn't working)
