### 2024-10-08 trchudley/pdemtools#23 @maiwinstrup
Plots are still not showing in my version of Jupyterlab (newest version, clean environment). But if you add plt.show(), as also suggested by Adam, everything works fine!
