### 2024-10-08 trchudley/pdemtools#21 @trchudley
Great, I'm glad we've diagnosed this at least!

Unfortunately, the preview is provided by the PGC so there is no option for a mask-less preview. I was going to suggest that the preview function/image could be a good place to provide some sort of `if all nan` check, but:

 - the resolution of the preview is much larger and I believe formed via nearest neighbour resampling, so there could be some valid data at 2 m resolution not present in the preview. I'm not quite sure who would be interested in data this patchy, but it wouldn't be strictly correct to raise an `all nan` warning based on the larger-resolution data.
 - users may indeed be interested in data without the bitmask. The bitmask isn't perfect (arguably too aggressive) and there are other potential ways of filtering blunders/poor-quality data.

On balance, my instinct is to trust users to be able to make 'mistakes' here, but keep them educated and informed of the possibility. Would you agree with the proposal to introduce more written material in the documentation on the possibility of this happening at the appropriate locations in the notebooks and FAQ?
