### 2024-10-08 trchudley/pdemtools#23 @trchudley
Yep, great - I think this is all relict of me normally using `%matplotlib widget` outside of training notebooks! I'll going to add plt.show() to everything as recommended in #33. Apologies and thanks for flagging!
