### 2024-10-08 40ec981
remove indices and tables from homepage #25
