### 2024-10-08 4b20b24
notebooks use plt.show() #23 #33
