### 2024-10-08 trchudley/pdemtools#23 @trchudley
`plt.show()` now standard throughout the notebooks. This will now be visible in the 'latest' version of the readthedocs (https://pdemtools.readthedocs.io/en/latest/) - NB following #32 website will default to stable version.
