### 2024-10-08 trchudley/pdemtools#22 @maiwinstrup
I tested the updated version (main branch) running the 'strip_search_and_dem_differencing' notebook, but the gdf returned in box 3 contains a different set of keys than the ones mentioned above?
