### 2024-10-08 trchudley/pdemtools#21 @maiwinstrup
Yes, I think that would be fine to just add some explanation to make the user aware that this may happen, and under which circumstances.
