### 2024-10-08 trchudley/pdemtools#17 @maiwinstrup
Great! I have read the new text on Supplementary datafiles, which clarifies the above issues.
