### 2024-10-08 trchudley/pdemtools#22 @trchudley
Thanks Mai. I'm struggling to replicate this - from my own main branch install if I print the returned `gdf` columns I get the following:

```python
gdf.columns
```
```bash
Index(['fid', 'dem_id', 'pairname', 'stripdemid', 'sensor1', 'sensor2',
       'catalogid1', 'catalogid2', 'acqdate1', 'acqdate2', 'gsd', 'epsg',
       'setsm_ver', 's2s_ver', 'creationdate', 'geocell', 'is_lsf',
       'is_xtrack', 'valid_area_matchtag_density', 'valid_area_sqkm',
       'valid_area_percent', 'water_area_sqkm', 'water_area_percent',
       'cloud_area_sqkm', 'cloud_area_percent', 'avg_convergence_angle',
       'avg_expected_height_accuracy', 'avg_sun_elev1', 'avg_sun_elev2',
       'rmse', 'fileurl', 's3url', 'geom', 'pdt_time1', 'pdt_time2',
       'pdt_dem_baseline_hours', 'pdt_time_mean', 'pdt_year', 'pdt_month',
       'pdt_aoi_frac'],
      dtype='object')
```

Note the relevant columns (`'pdt_time1', 'pdt_time2', 'pdt_dem_baseline_hours', 'pdt_time_mean', 'pdt_year', 'pdt_month', 'pdt_aoi_frac'`) concatenated to the end of the notebook, which is the expected behaviour of the new default `include_custom_columns = True` behaviour.

Is something different happening for you? Can we confirm with `help(pdt.search)` that the `include_custom_columns` option is present?
