### 2024-10-08 d2f3025
docs clarification on all-nan scenes
