### 2024-10-08 trchudley/pdemtools#21 @trchudley
The following explanation has been added to the notebook at https://pdemtools.readthedocs.io/en/latest/examples/strip_search_and_dem_difference.html

> **NOTE**: The `preview()` option draws on PGC-provided data, which is automatically masked according to the PGC bitmask. It is entirely possible that the `search()` function can return a valid datastrip that covers a sufficient proportion of the AOI to meet the `min_aoi_frac` requirements, but it will appear empty (i.e. all-`NaN`) when viewed. The `preview()` function will help you identify these 'empty' scenes, but there may still be (poor-quality) data present if it is downloaded using the `load.from_search()` function with `bitmask=False`. However, by default, `load.from_search()`defaults to `bitmask = False`.

And the following added to the FAQ at https://pdemtools.readthedocs.io/en/latest/appendix/faq.html

>> Q: I have downloaded a strip scene, but the strip appears to have no data within the area. Why?

> Strips can appear empty (i.e. `NaN` values) when photogrammetry fails (due to cloud, water, etc) or it is masked by the PGC bitmask. Data previewed by the `pdt.load.preview()` function is masked according to this bitmask, and data loaded by the `pdt.load.from_search()` function is also masked by default (this can be disabled by using the `bitmask = False` option).

> As a result, it is entirely possible that the `search()` function can return a valid datastrip that covers a sufficient proportion of the AOI to meet the `min_aoi_frac` requirements, but it will appear empty (i.e. all-`NaN`) when viewed. The `preview()` function will help you identify these 'empty' scenes, but there may still be (poor-quality) data present if it is downloaded using the `load.from_search()` function with `bitmask = False`.

(please allow some time for changes to filter through to the public site!)
