### 2024-10-09 trchudley/pdemtools#22 @maiwinstrup
Also - one last thing: I find it confusing that both dates and (months, years) are provided as input to the search function in the "strip search" notebook. I suspect that the ‘dates’ can simply be removed?
