### 2024-10-09 trchudley/pdemtools#22 @trchudley
Hmm, interesting. Perhaps an obvious sanity check, but was the editable flag used when installing the dev version? (i.e. `pip install -e  .` from the root directory). If not, changes to the directory won't be reflected in the installed pip version. Might be worth uninstalling and reinstalling just in case.

> Also - one last thing: I find it confusing that both dates and (months, years) are provided as input to the search function in the "strip search" notebook. I suspect that the ‘dates’ can simply be removed?

You're correct that it would be redundant to include them all in a 'real' application, but considering that the notebook will be the primary way that users learn this process, I think it is useful to show the possible ways in which users might interact with the function. If it's all right, I'll instead highlight this in the notebook.
