### 2024-10-09 trchudley/pdemtools#22 @maiwinstrup
I installed it by pulling the main branch directly from the github repo? But, yes, it may just be some weird issue on my part. And, yes, fine to simply high-light the redundancy in the text.
