### 2024-10-09 trchudley/pdemtools#22 @trchudley
> I installed it by pulling the main branch directly from the github repo?

Thanks - yep, this sounds good. I assume this would then be installed via pip as per the [developer install](https://pdemtools.readthedocs.io/en/latest/getting_started/install.html). However, if the editable flag (`-e`) was missed out, your pip install won't automatically reflect any changes made when you pulled the updated repo. I would recommend uninstalling and reinstalling from the repo just to make sure (i.e. `pip uninstall pdemtools` > `pip instell -e .`). Then we can diagnose further...

> And, yes, fine to simply high-light the redundancy in the text.

Great, I've added the following:

> Note that including both `dates` and `years` as parameters in this use case is strictly redundant (because e.g. `dates = '20190101/20191231'` is functionally identical to `years = 2019`). However, it is worth highlighting that both options are avaiable to you! .
