+++ paper/paper.md
@@ -42,9 +42,9 @@ bibliography: paper.bib
 
 As Earth Science has moved into the 'big data' era, increasing amounts of Arctic- and Antarctic-focused resources are available as public, cloud-optimised datasets. New approaches are providing Python tools to act as combined API and processing tools, such as `icepyx` [@scheick_icepyx_2023] or `pypromice` [@how_pypromice_2023]. From 2022 (ArcticDEM v4.1 and REMA v2), the PGC DEM products are [hosted](https://polargeospatialcenter.github.io/stac-browser/#/external/pgc-opendata-dems.s3.us-west-2.amazonaws.com/pgc-data-stac.json) as Cloud Optimised GeoTIFFs (CoGs) in a SpatioTemporal Asset Catalog (STAC), a standardised structure for cataloguing spatiotemporal data. However, the PGC STAC is currently a 'static' rather than 'dynamic' STAC, which means there is no convenient Application Programming Interface (API) for searching the datasets in response to user queries. This limits the ability of users to programmatically interact with ArcticDEM and REMA data in a quick and efficient manner. The `pdemtools` package has two aims: the first is to provide a Python-focussed alternative for searching and downloading ArcticDEM and REMA data, emulating dynamic STAC query tools such as `pystac` [@radiant_pystac_2024]; whilst the second is to provide commonly used processing functions specific to the needs of ArcticDEM and REMA users (a focus on ice sheet and cryosphere work), as well as the particular strengths of the ArcticDEM and REMA datasets (high-resolution and multitemporal).
 
-The `pdemtools` `search()` tool and `load` module allow for convenient access to the ArcticDEM and REMA datasets. Mosaics can be downloaded from a one-line `load.mosaic()` function, whilst the `search()` function allows for convenient filtering of a locally downloading ArcticDEM/REMA strip index according to variables such as date, region of interest, spatial coverage, temporal baseline, source sensors, accuracy, and cross-track data. The results of searches are returned as a `geopandas` [@jordahl_geopandas_2024] dataframe, and can be downloaded using the `load.from_search()` function. Elevation models are returned as `xarray` DataArrays [@hoyer_xarray_2017] with geospatial metadata via the `rioxarray` extension [@corteva_rioxarray_2024] - a standard format for storing and processing n-dimensional geospatial data within the geospatial Python community. By utilising standardised formats, the aim is to allow the user to quickly move beyond `pdemtools` into their own analysis in whatever format they desire, be that `xarray`, `numpy` or `dask` datasets, DEM analysis Python packages such as `xdem` [@glaciohack_xdem_2020] for advanced coregistration or `richdem` [@barnes_richdem_2016] for flow analysis, or exporting to geospatial file formats for analysis beyond Python.
[…]
