+++ paper/paper.md
@@ -6,8 +6,8 @@ tags:
   - remote-sensing
   - geomorphometry
   - glaciology
-  - greenland
-  - antarctica
+  - Greenland
+  - Antarctica
   - arctic
 authors:
   - name: Thomas R. Chudley
@@ -38,25 +38,25 @@ bibliography: paper.bib
 
 # Statement of need
 
-[ArcticDEM](https://www.pgc.umn.edu/data/arcticdem/) and [REMA](https://www.pgc.umn.edu/data/rema/) are high-resolution, time-stamped 2-metre-resolution DEMs of the polar regions provided by the Polar Geospatial Center (PGC). They are extracted by applying stereo auto-correlation techniques [@noh_surface_2017] to pairs of submetre Maxar satellite imagery, including Worldview-1, Worldview-2, Worldview-3, and GeoEye-1, beginning in 2007 (ArcticDEM) or 2009 (REMA) and ongoing to the present day. Products are available as tens of thousands of time-stamped 'strips' [@porter_arcticdem_2022; @howat_remastrips_2022] constructed from individual scene-pairs, or as a single mosaic [@porter_arcticdem_2023; @howat_remamosaic_2022] compiled from the combined stack of strips. Strips allow users to perform change detection by comparing data from different seasons or years, whilst mosaics provide a consistent and comprehensive product over the entire polar regions.
+[ArcticDEM](https://www.pgc.umn.edu/data/arcticdem/) and [REMA](https://www.pgc.umn.edu/data/rema/) are high-resolution, time-stamped 2-metre-resolution DEMs of the polar regions provided by the Polar Geospatial Center (PGC). They are extracted by applying stereo auto-correlation techniques [@noh_surface_2017] to pairs of submetre Maxar satellite imagery. The data includes Worldview-1, Worldview-2, Worldview-3, and GeoEye-1, beginning in 2007 (ArcticDEM) or 2009 (REMA) and ongoing to the present day. Products are available as tens of thousands of time-stamped 'strips' [@porter_arcticdem_2022; @howat_remastrips_2022] constructed from individual scene pairs, or as a single mosaic [@porter_arcticdem_2023; @howat_remamosaic_2022] compiled from the combined stack of strips. Strips allow users to perform change detection by comparing data from different seasons or years, whilst mosaics provide a consistent and comprehensive product over the entire polar regions.
 
[…]
