### 2024-10-14 trchudley/pdemtools#38 @AdamRJensen
Python packaging
The packaging currently relies on a conf.py file which is a bit of an outdated practice.

I highly suggest switching to using a pyproject.toml and setup.cfg files which is the current recommended practice: https://packaging.python.org/en/latest/tutorials/packaging-projects/#configuring-metadata

You can check out the twoaxistracking package as inspiration: [setup.cfg](https://github.com/pvlib/twoaxistracking/blob/main/setup.cfg) and [pyproject.toml](https://github.com/pvlib/twoaxistracking/blob/main/pyproject.toml). Note how the versions of the documentation packages are all fixes (this is recommended to avoid your documentation breaking in the future as new releases become available).

I'm partly recommending this because when I tried to build the documentation, then my environment didn't have Sphinx for example. If using the above approach, one can then get all the documentation packages by installing as:

    pip install pydem[doc]
