### 2024-10-14 trchudley/pdemtools#38 @AdamRJensen
The second `[environment.yml](https://github.com/trchudley/pdemtools/blob/main/docs/environment.yml)` file in the docs folder threw me off.

My recommendation would be to get rid of both of the environment.yml files and include the dependencies in the pyproject.toml as described here: https://packaging.python.org/en/latest/guides/writing-pyproject-toml/
