### 2024-10-15 trchudley/pdemtools#38 @AdamRJensen
Here's a bit more info:

> Starting with [PEP 621](https://peps.python.org/pep-0621/), the Python community selected pyproject.toml as a standard way of specifying project metadata. Setuptools has adopted this standard and will use the information contained in this file as an input in the build process.

[Source](https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html)
