### 2024-10-15 56daa9a
add statement of need to docs #37
