+++ paper/paper.bib
@@ -53,17 +53,17 @@
 	pages = {5298},
 }
 
-@misc{howat_remamosaic_2022,
-	title = {The {Reference} {Elevation} {Model} of {Antarctica} - {Mosaics}, {Version} 2},
-	doi = {10.7910/DVN/EBW8UC},
+@misc{howat_remastrips_2022,
+	title = {The {Reference} {Elevation} {Model} of {Antarctica} - {Strips}, {Version} 4.1},
+	doi = {10.7910/DVN/X7NDNY},
 	publisher = {Harvard Dataverse},
 	author = {Howat, Ian and Porter, Claire and Noh, Myoung-Jon and Husby, Erik and Khuvis, Samuel and Danish, Evan and Tomko, Karen and Gardiner, Judith and Negrete, Adelaide and Yadav, Bidhyananda and Klassen, James and Kelleher, Cole and Cloutier, Michael and Bakker, Jesse and Enos, Jeremy and Arnold, Galen and Bauer, Greg and Morin, Paul},
 	year = {2022},
 }
 
-@misc{howat_remastrips_2022,
-	title = {The {Reference} {Elevation} {Model} of {Antarctica} - {Strips}, {Version} 4.1},
-	doi = {10.7910/DVN/X7NDNY},
+@misc{howat_remamosaic_2022,
+	title = {The {Reference} {Elevation} {Model} of {Antarctica} - {Mosaics}, {Version} 2},
+	doi = {10.7910/DVN/EBW8UC},
 	publisher = {Harvard Dataverse},
 	author = {Howat, Ian and Porter, Claire and Noh, Myoung-Jon and Husby, Erik and Khuvis, Samuel and Danish, Evan and Tomko, Karen and Gardiner, Judith and Negrete, Adelaide and Yadav, Bidhyananda and Klassen, James and Kelleher, Cole and Cloutier, Michael and Bakker, Jesse and Enos, Jeremy and Arnold, Galen and Bauer, Greg and Morin, Paul},
 	year = {2022},
