+++ docs/appendix/references.md
@@ -2,9 +2,13 @@
 
 Florinsky, I. V. (2009). Computation of the third‐order partial derivatives from a digital elevation model. _International journal of geographical information science_, 23(2), 213-231. https://doi.org/10.1080/13658810802527499
 
+Foerste, C., _et al._ (2014). EIGEN-6C4 the latest combined global gravity field model including GOCE data up to degree and order 2190 of GFZ potsdam and GRGS toulouse. GFZ Data Services. https://doi.org/10.5880/icgem.2015.1
+
 Howat, I., _et al._ (2022a). The Reference Elevation Model of Antarctica – Strips, Version 4.1. _Harvard Dataverse_ https://doi.org/10.7910/DVN/X7NDNY
 
 Howat, I., _et al._ (2022b). The Reference Elevation Model of Antarctica – Mosaics, Version 2, _Harvard Dataverse_ https://doi.org/10.7910/DVN/EBW8UC
+
+Hoyer, S., & Hamman, J. (2017). xarray: N-d labeled Arrays and Datasets in Python. _Journal of Open Research Software_, 5(1), 10. https://doi.org/10.5334/jors.148
 
 MacGregror, J. A. _et al._ (2024). Geologic Provinces Beneath the Greenland Ice Sheet Constrained by Geophysical Data Synthesis. _Geophysical Research Letters_, 51, e2023GL107357. https://doi.org/10.1029/2023GL107357
 
@@ -14,12 +18,16 @@ Morlighem, M. _et al._ (2022a). IceBridge BedMachine Greenland, Version 5 [Data 
 
 Morlighem, M. _et al._ (2022b). MEaSUREs BedMachine Antarctica, Version 3 [Data Set]. _NASA National Snow and Ice Data Center Distributed Active Archive Center_. https://doi.org/10.5067/FPSU0V1MWUB6
 
+Noh, M. J., & Howat, I. M. (2017). The surface extraction from TIN based search-space minimization (SETSM) algorithm. _ISPRS Journal of Photogrammetry and Remote Sensing_, 129, 55-76. https://doi.org/10.1016/j.isprsjprs.2017.04.019
+
 Nuth, C. and Kääb, A. (2011) Co-registration and bias corrections of satellite elevation data sets for quantifying glacier thickness change, _The Cryosphere_, 5, 271–290, https://doi.org/10.5194/tc-5-271-2011
 
 Porter, C., _et al._ (2022). ArcticDEM - Strips, Version 4.1. _Harvard Dataverse_. https://doi.org/10.7910/DVN/OHHUKH
 
 Porter, C., _et al._ (2023), ArcticDEM, Version 4.1, _Harvard Dataverse_. https://doi.org/10.7910/DVN/3VDC4W
 
+Scheick et al., (2023). icepyx: querying, obtaining, analyzing, and manipulating ICESat-2 datasets. _Journal of Open Source Software_, 8(84), 4912, https://doi.org/10.21105/joss.04912
+
 Shiggins, C. J., _et al._ (2023). Automated ArcticDEM iceberg detection tool: insights into area and volume distributions, and their potential application to satellite imagery and modelling of glacier–iceberg–ocean systems, _The Cryosphere_, 17, 15–32, https://doi.org/10.5194/tc-17-15-2023
 
-Zevenbergen, L. W. and Thorne, C. R. (1987). Quantitative analysis of land surface topography. Earth surface processes and landforms, 12(1), 47-56.
+Zevenbergen, L. W. and Thorne, C. R. (1987). Quantitative analysis of land surface topography. Earth surface processes and landforms, 12(1), 47-56. https://doi.org/10.1002/esp.3290120107
