+++ docs/getting_started/why_pdemtools.md
@@ -1,0 +1,13 @@
+# Why pDEMtools?
+
+> This page reproduces the 'Statement of Need' of the [_Journal of Open Source Software_ manuscript](https://joss.theoj.org/papers/2a10e67d2709a6cfb672538b4a21726e) for `pdemtools`.
+
+[ArcticDEM](https://www.pgc.umn.edu/data/arcticdem/) and [REMA](https://www.pgc.umn.edu/data/rema/) are high-resolution, time-stamped 2-metre-resolution DEMs of the polar regions provided by the Polar Geospatial Center (PGC). They are extracted by applying stereo auto-correlation techniques ([Noh & Howat, 2017](https://doi.org/10.1016/j.isprsjprs.2017.04.019)) to pairs of submetre Maxar satellite imagery. The data includes Worldview-1, Worldview-2, Worldview-3, and GeoEye-1, beginning in 2007 (ArcticDEM) or 2009 (REMA) and ongoing to the present day. Products are available as tens of thousands of time-stamped 'strips' ([Porter _et al._ 2022](ttps://doi.org/10.7910/DVN/OHHUKH); [Howat _et al._ 2022a](https://doi.org/10.7910/DVN/X7NDNY)) constructed from individual scene pairs, or as a single mosaic ([Porter _et al._ 2023](https://doi.org/10.7910/DVN/3VDC4W); [Howat _et al._ 2022a](https://doi.org/10.7910/DVN/EBW8UC)) compiled from the combined stack of strips. Strips allow users to perform change detection by comparing data from different seasons or years, whilst mosaics provide a consistent and comprehensive product over the entire polar regions.
+
+As Earth Science has moved into the 'big data' era, increasing amounts of Arctic- and Antarctic-focused resources are available as public, cloud-optimised datasets. New approaches are providing Python tools to act as combined API and processing tools, such as [`icepyx`](https://icepyx.readthedocs.io) ([Scheick _et al._ 2023](https://doi.org/10.21105/joss.04912)) or [`pypromice`](https://pypromice.readthedocs.io) ([How _et al._ 2023](https://doi.org/10.21105/joss.05298)). From 2022 (ArcticDEM v4.1 and REMA v2), the PGC DEM products are [hosted](https://polargeospatialcenter.github.io/stac-browser/#/external/pgc-opendata-dems.s3.us-west-2.amazonaws.com/pgc-data-stac.json) as Cloud Optimised GeoTIFFs (CoGs) in a SpatioTemporal Asset Catalog (STAC), a standardised structure for cataloguing spatiotemporal data. However, the PGC STAC is currently a 'static' rather than 'dynamic' STAC, which means there is no convenient Application Programming Interface (API) for searching the datasets in response to user queries. This limits the ability of users to programmatically interact with ArcticDEM and REMA data in a quick and efficient manner. The `pdemtools` package has two aims: the first is to provide a Python-focussed alternative for searching and downloading ArcticDEM and REMA data, emulating dynamic STAC query tools such as [`pystac`](https://pystac.readthedocs.io); whilst the second is to provide commonly used processing functions specific to the needs of ArcticDEM and REMA users (a focus on ice sheet and cryosphere work), as well as the particular strengths of the ArcticDEM and REMA datasets (high-resolution and multitemporal).
+
[…]
