### 2024-10-15 1dd218e
add mathjax to docs
