### 2024-10-15 704f3f0
add mathjax to docs
