+++ docs/getting_started/why_pdemtools.md
@@ -8,6 +8,7 @@ As Earth Science has moved into the 'big data' era, increasing amounts of Arctic
 
 The `pdemtools` `search()` tool and `load` module allow for convenient access to the ArcticDEM and REMA datasets. Mosaics can be downloaded from a one-line `load.mosaic()` function, whilst the `search()` function allows for convenient filtering of a locally downloading ArcticDEM/REMA strip index according to variables such as date, region of interest, spatial coverage, temporal baseline, source sensors, accuracy, and cross-track data. The results of searches are returned as a [`geopandas`](https://geopandas.org) dataframe, and can be downloaded using the `load.from_search()` function. Elevation models are returned as [`xarray`](https://docs.xarray.dev) DataArrays [Hoyer _et al. 2017](https://doi.org/10.5334/jors.148) with geospatial metadata via the [`rioxarray`](https://corteva.github.io) extension - a standard format for storing and processing n-dimensional geospatial data within the geospatial Python community. By utilising standardised formats, the aim is to allow the user to quickly move beyond `pdemtools` into their own analysis in whatever format they desire, be that `xarray`, `numpy` or `dask` datasets, DEM analysis Python packages such as [`xdem`](https://xdem.readthedocs.io/) for advanced coregistration or [`richdem`](https://richdem.readthedocs.io) for flow analysis, or exporting to geospatial file formats for analysis beyond Python.
 
-After download, there exist a number of (pre-)processing steps that are near universally common in topographic analyses. These include geoid-correction, co-registration of time-series data, and/or the construction of terrain parameters such as hillshade, slope, aspect, and curvature. `pdemtools` contains pre-built functions to perform these processing steps, as well as further functionality specific to ArcticDEM and REMA use cases. For instance, we include functions to quickly extract the EIGEN-6C4 geoid [Foerste _et al._ 2014](https://doi.org/10.5880/icgem.2015.1) and Greenland/Antarctic bedrock masks directly from local versions of the Greenland and Antarctic BedMachine datasets ([Morlighem _et al_, [2022a](https://doi.org/10.5067/GMEVBWFLWA7X), [2022b](https://doi.org/10.5067/FPSU0V1MWUB6)), reprojecting and resampling the data to match the target DEM. Options for ingesting user-provided mask and geoid data are also provided. Additionally, partial derivatives of the surface used to calculate terrain parameters ($\frac{\partial z}{\partial x}$, $\frac{\partial z}{\partial y}$, $\frac{\partial^2 z}{\partial x^2}$, $\frac{\partial^2 z}{\partial y^2}$, $\frac{\partial^2 z}{\partial x \partial y}$) are calculated following [Florinsky (2009)](https://doi.org/10.1080/13658810802527499), as opposed to more common methods such as [Zevenbergen and Thorne (1987)](https://doi.org/10.1002/esp.3290120107). The newer approach computes partial derivatives of elevation based on fitting a third-order polynomial, by the least-squares approach, to a 5 $\times$ 5 window as opposed to the more common 3 $\times$ 3 window. This is more appropriate for high-resolution DEMs: curvature over a 10 m window for the 2 m resolution ArcticDEM/REMA strips will lead to a local denoising effect that limits the impact of noise common in high-resolution photogrammetric products. These methods are also adapted into a co-registration routine, which otherwise follows the commonly used approach of [Nuth & Kääb (2011)](https://doi.org/10.5194/tc-5-271-2011).
[…]
