+++ docs/notes.txt
@@ -1,16 +1,0 @@
-To build the website using sphinx,
-
-cd docs
-make html
-
-In a pinch, 'make clean' might fix things.
-
-When you're happy, push to git (incl git add . if you've got new files!)
-
-Go to readthedocs and create a new build
-
-To think about:
- - Automatic versioning?
-
-Pages:
-https://lpn-doc-sphinx-primer-devel.readthedocs.io/
