### 2024-10-15 trchudley/pdemtools#38 @trchudley
Thanks Adam. Yes, `pdemtools` itself is using a `pyproject.toml` file. However, the readthedocs/sphinx website is built on top of a sample repo that used a `conf.py` file. I've attempted to look into this and it doesn't look like sphinx fully/consistently supports `pyproject.toml` files (yet?). The official documentation [still describes `conf.toml` files](https://www.sphinx-doc.org/en/master/usage/configuration.html). Our chosen website theme also still recommends a `conf.py` file in the [setup tutorial](https://sphinx-book-theme.readthedocs.io/en/stable/tutorials/get-started.html). I don't think this is simply a case of outdated docs - even 'big' projects like `xarray` [are still using `conf.py` files in their docs](https://github.com/pydata/xarray/tree/main/doc). The [`sphinx-pyproject`](https://sphinx-pyproject.readthedocs.io) project, which claims to move "some" (?) of the sphinx configuration into a `pyproject.toml`, is still in development mode. I guess some non-Python sphinx background processes in the makefile may depend on the `conf.py` file in an unusual manner?

I am very much an end-user of the sphinx/readthedocs tools and getting into the weeds of the sphinx configuration is making me a little anxious about breaking things. If it's all right can I defer moving the website to a `toml` file until Sphinx community development and tutorials/support surrounding `pyproject.toml` is more mature?

As an interim, I have included a [`README.md` file in the `docs` directory](https://github.com/trchudley/pdemtools/blob/main/docs/README.md) that outlines how users can edit and build the website (although I very few outside the core `pdemtools` team should really need to). Is this acceptable to you?
