### 2024-10-15 4c859c8
plaintext partial derivatives in docs
