### 2024-10-15 39208f5
fix typo in why_pdemtools.md docs
