### 2024-10-15 #7149 @trchudley
Hi @AdamRJensen, I've responded to your last few issues. If you're OK with them (particularly my response to https://github.com/trchudley/pdemtools/issues/38), please let me know here and I will release the post-review version (0.8.5), archive the release on zenodo, and post the link here.
