+++ docs/environment.yml
@@ -1,24 +1,0 @@
-name: pdemtools_env
-channels:
-  - conda-forge
-  - defaults
-dependencies:
-  - python=>3.10
-  - rioxarray
-  - rasterio
-  - geopandas 
-  - pandas
-  - opencv 
-  - scipy 
-  - numba 
-  - gdal
-  - pandas
-  - numpy
-  - shapely
-  - pyarrow
-  - sphinx==7.2.6
-  - myst-parser
-  - nbsphinx
-  - nbsphinx-link
-  # - sphinx-rtd-theme
-  - sphinx-book-theme
