+++ environment.yml
@@ -1,17 +1,0 @@
-name: pdemtools_env
-channels:
-  - conda-forge
-  - defaults
-dependencies:
-  - python>=3.10
-  - rioxarray 
-  - geopandas 
-  - opencv 
-  - scipy 
-  - numba 
-  - gdal
-  - pandas
-  - numpy
-  - shapely
-  - pyarrow
-  - fiona
