### 2024-10-15 c0c90ac
Add dependencies to pyproject.toml
