+++ pyproject.toml
@@ -9,7 +9,30 @@ description = "Conveniently search, download, and preprocess ArcticDEM and REMA 
 keywords = [ "geospatial", "elevation", "arcticdem", "rema", "dem",]
 classifiers = [ "Intended Audience :: Science/Research", "Programming Language :: Python :: 3", "License :: OSI Approved :: MIT License", "Operating System :: OS Independent", "Topic :: Scientific/Engineering", "Topic :: Scientific/Engineering :: GIS", "Topic :: Scientific/Engineering :: Image Processing",]
 requires-python = ">=3.10"
-dependencies = ["rioxarray", "rasterio", "geopandas", "pandas", "shapely", "numpy", "GDAL", "opencv-python", "scipy", "numba", "pyarrow", "fiona"]  # GDAL
+dependencies = [
+    "rioxarray",
+    "rasterio",
+    "geopandas",
+    "pandas",
+    "shapely",
+    "numpy",
+    "GDAL",
+    "opencv-python",
+    "scipy",
+    "numba",
+    "pyarrow",
+    "fiona",
+]  # GDAL
+
+[project.optional-dependencies]
+test = ["pytest"]
+doc = [
+    "sphinx==7.2.6",
+    "myst-parser",
+    "nbsphinx",
+    "nbsphinx-link",
+    "sphinx-book-theme",
+]
 
 [[project.authors]]
 name = "Tom Chudley"
