### 2024-10-15 trchudley/pdemtools#38 @AdamRJensen
@trchudley Sorry for the poor explanation of this issue. I was not concerned with the conf.py for the docs, but initially I thought you were using a conf.py for the packaging (though this is not the case). The changes in #39 is really what I meant, i.e., removing the two conda environment files if possible.
