### 2024-10-15 trchudley/pdemtools#38 @trchudley
Thanks Adam. I've had a look at the pull request.

I definitely wouldn't want to remove the root directory yaml file. This is necessary to set up our github actions ([e.g. testing](https://github.com/trchudley/pdemtools/blob/main/.github/workflows/unit_test.yml)), and is the [recommended way of installing the dev environment](https://pdemtools.readthedocs.io/en/latest/getting_started/install.html#developer-install). This is so that we can enforce installation via conda, given our previously discussed issues surrounding GDAL and pip. A base environment yml file appears to be common in peer projects for dev install purposes (e.g. [Penny How's pypromice](https://github.com/GEUS-Glaciology-and-Climate/pypromice), also a JOSS paper, recommends [dev installation via the yml file](https://pypromice.readthedocs.io/en/latest/install.html#developer-install)).

As for the docs yml file, this is for the same reason: so that we can enforce readthedocs to create the website within a GDAL-friendly conda environment (hence the inclusion of the `conda: environment: docs/environment.yml` lines that you recommended deleting). Whilst it looks like introducing pdemtools[doc] optional dependencies would allow us to at least synthesise this into the root directory yaml file, I am not aware of a way for readthedocs to infer an optional dependency install: this is the reason for the seperate `docs/environment.yml` file.

I'm open to other ideas if I've missed anything. Perhaps it might be sufficient to rename `docs/environment.yml` to `doc/readthedocs_env.yml` or similar to avoid confusion?
