+++ docs/readthedocs_conda_env.yml
@@ -17,8 +17,7 @@ dependencies:
   - shapely
   - pyarrow
   - sphinx==7.2.6
-  - myst-parser
-  - nbsphinx
-  - nbsphinx-link
-  # - sphinx-rtd-theme
-  - sphinx-book-theme
+  - myst-parser==4.0.0
+  - nbsphinx==0.9.5
+  - nbsphinx-link==1.3.0
+  - sphinx-book-theme==1.1.3
