### 2024-10-15 #7149 @AdamRJensen
@trchudley Alright, we've made it! Thank you for your patience 🥳

Please go ahead and make a new release and upload it to Zenodo and write back with the version number and DOI
