### 2024-10-15 #7149 @trchudley
Thanks @AdamRJensen for your time!

v0.8.5 of `pdemtools` has been released and is archived on Zenodo at https://doi.org/10.5281/zenodo.13936813
