### 2024-10-15 #7149 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
