### 2024-10-22 #7149 @kthyng
Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file (such as "toulouse" but please check for others)
