### 2024-10-22 fc73032
capitalise bib file
