+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @misc{barnes_richdem_2016,
-    title = {RichDEM: Terrain Analysis Software}, 
+    title = {{RichDEM}: {Terrain} {Analysis} {Software}}, 
     author = {Barnes, Richard }, 
     year = {2016 - 2022}, 
     version = {2.3.1},
@@ -7,7 +7,7 @@
 }
 
 @misc{corteva_rioxarray_2024,
-    author = {Corteva },
+    author = {Corteva},
     title = {rioxarray: geospatial xarray extension powered by rasterio},
     year = {2019 - 2024},
     version = {0.15.5},
@@ -27,7 +27,7 @@
 }
 
 @misc{foerste_geoid_2014,
-	title = {EIGEN-6C4 The latest combined global gravity field model including GOCE data up to degree and order 2190 of GFZ Potsdam and GRGS Toulouse},
+	title = {{EIGEN-6C4} {The} latest combined global gravity field model including {GOCE} data up to degree and order 2190 of {GFZ} {Potsdam} and {GRGS} {Toulouse}},
 	doi = {10.5880/icgem.2015.1},
 	publisher = {GFZ Data Services},
 	author = {Foerste, Christoph and Bruinsma, Sean. L. and Abrykosov, Oleh and Lemoine, Jean-Michel and Marty, Jean Charles and Flechtner, Frank and Balmino, Georges and Barthelmes, Franz and Biancale, Richard},
@@ -35,7 +35,7 @@
 }
 
 @misc{glaciohack_xdem_2020,
-    title = {xDEM: Analysis of digital elevation models}, 
+    title = {{xDEM}: {Analysis} of digital elevation models}, 
     author = {{GlacioHack Community} }, 
     year = {2020 - 2023}, 
 	version = {0.0.18},
