### 2024-10-22 #7149 @trchudley
Thanks @kthyng, I hope you enjoyed your travels! I think I've fixed up the missing capitalisation. I've also moved my first definition of Application Programming Interface to slightly earlier in the relevant paragraph - I realised that I used 'API' once without defining.
