### 2024-10-23 #7149 @kthyng
Congratulations on your new publication @trchudley! Many thanks to editor @AdamRJensen and to reviewers @mn5hk and @maiwinstrup (plus bonus review from @AdamRJensen!) for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
